Try 3 - Add Reserved BIOS Boot area partition
The current suspicion is that grub install fails without the Reserved BIOS Boot
area partition.

I created this small 1MB partition on each drive.  There is no raid across the
drives for this partition.  This attempt retries with RAID5 on the OS volume
group which contains the logical volume for / and swap.  The reason I did RAID5
was to double the usable space.